requirements.txt
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-anthropic>=0.1.1
llama-index>=0.10.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
pinecone-client>=3.0.0
sentence-transformers>=2.5.0

# Data Processing
pypdf2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
unstructured>=0.12.0
python-magic>=0.4.27

# Web & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0
gradio>=4.19.0
requests>=2.31.0
httpx>=0.26.0

# Database
sqlalchemy>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.6.0
tenacity>=8.2.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Monitoring & Logging
langsmith>=0.1.0
loguru>=0.7.0
prometheus-client>=0.20.0

# Testing & Evaluation
ragas>=0.1.0
deepeval>=0.20.0
